+++ bin/aidomx.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import { AidomxCLI } from '../dist/index.cjs'
+
+await AidomxCLI()