+++ src/commands/index.ts
@@ -4,11 +4,12 @@ Aidomx CLI
 Usage: 
   
 Commands:  
-aidomx  init     Inisialisasi proyek dengan konfigurasi AIDOMX
-aidomx  dev      Menjalan dev server untuk client
-aidomx  start    Menjalankan server dengan php atau node
-aidomx  version  Menampilkan versi CLI  
-aidomx  help     Menampilkan bantuan ini  
+aidomx  init      Inisialisasi proyek dengan konfigurasi AIDOMX
+aidomx  dev       Menjalan dev server untuk client
+aidomx  generate  Generate layout dan ui
+aidomx  start     Menjalankan server dengan php atau node
+aidomx  version   Menampilkan versi CLI  
+aidomx  help      Menampilkan bantuan ini  
 `
 
 export { Initialize } from './init'