+++ src/index.ts
@@ -1,13 +1,10 @@
+import { createNewCLI } from './cli'
 /**
- * Aidomx Entry File
+ * Aidomx CLI Entry
  *
- * @file src/index.ts
- * @created 2025-04-07
+ * @since 2025-04-07
  * @author aidomx
- * @project AIDOMX
  * @license MIT
- * @contact github.com/aidomx
+ * @see github.com/aidomx
  */
-export type * from './types'
-export * from './commands'
-export { swe } from './schemas'
+export const AidomxCLI = async () => await createNewCLI()