+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export type * from './cli'
+export type * from './engine'
 export type * from './server'
 export type * from './schemas'